moving utility functions. Nothing to see here.
