+++ lib/widget/chart_highcharts_builder.dart
@@ -1,5 +1,6 @@
 import 'package:ensemble/framework/widget/widget.dart';
 import 'package:ensemble/util/utils.dart';
+import 'package:ensemble/widget/helpers/controllers.dart';
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
 import 'package:high_chart/high_chart.dart';
 import 'package:flutter/material.dart';
