+++ lib/widget/ensemble_icon.dart
@@ -5,6 +5,7 @@ import 'package:ensemble/framework/widget/icon.dart' as ensembleLib;
 import 'package:ensemble/screen_controller.dart';
 import 'package:ensemble/util/utils.dart';
 import 'package:ensemble/framework/widget/widget.dart';
+import 'package:ensemble/widget/helpers/controllers.dart';
 import 'package:flutter/material.dart';
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
