+++ lib/widget/fintech/finicityconnect/finicityconnect.dart
@@ -3,6 +3,7 @@ import 'package:ensemble/framework/action.dart';
 import 'package:ensemble/framework/widget/widget.dart';
 import 'package:ensemble/screen_controller.dart';
 import 'package:ensemble/util/utils.dart';
+import 'package:ensemble/widget/helpers/controllers.dart';
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
 import 'package:flutter/material.dart';
 import 'package:ensemble/widget/fintech/finicityconnect/finicityconnectstate.dart';
