+++ lib/widget/markdown.dart
@@ -5,6 +5,7 @@ import 'package:ensemble/framework/error_handling.dart';
 import 'package:ensemble/screen_controller.dart';
 import 'package:ensemble/util/utils.dart';
 import 'package:ensemble/framework/widget/widget.dart' as framework;
+import 'package:ensemble/widget/helpers/controllers.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:flutter/material.dart';
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
@@ -49,7 +50,7 @@ class Markdown extends StatefulWidget with Invokable, HasController<MarkdownCont
 }
 
 
-class MarkdownController extends framework.WidgetController {
+class MarkdownController extends WidgetController {
   String? text;
 
   TextStyle? textStyle;
