+++ lib/widget/visualization/diagram.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:ensemble/framework/widget/widget.dart';
+import 'package:ensemble/widget/helpers/controllers.dart';
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
 import 'package:flutter/material.dart';
 import 'dart:ui' as ui;
